Сделал scaffold для базы деканат Оптимизировал получение профилей для вывода основной таблицы

+++ Infrastructure/Services/Interface/ISearchEntityService.cs
@@ -4,7 +4,7 @@ namespace Ifrastructure.Interface
 {
     public interface ISearchEntityService
     {
-        public Task<short?> SearchEdukind(string text);
+        public Task<int?> SearchEdukind(string text);
         public Task<int?> SearchCaseSDepartment(string text);
         public Task<StatusDiscipline> SearchStatusDiscipline(string text);
         public Task<LevelEdu> SearchLevelEdu(string text);
